resgain_django/requirements.txt
Django==2.1.3
PyMySQL==0.9.2
djangorestframework==3.9.0